brings back git utils and tests
